+++ src/potential_user/views.py
@@ -4,7 +4,7 @@ from django.views.generic import CreateView
 from potential_user.forms import RegistrationForm
 from potential_user.models import ApplicationForm
 from potential_user.utils import get_telegram_id
-from core.views import send_registration_email
+from core.utils import send_registration_email
 
 
 class RegistrationCreateView(CreateView):
